화폐 오버플로우 해결
double 형 자료를 넘어서는 인플레가 존재할까요..

+++ Assets/MainGameBackUp/Scripts/GameController.cs
@@ -254,7 +254,7 @@ public class GameController : MonoBehaviour
 		}
 
 		IEnumerable<MergePreset> presets = MergeController.Instance.GetPresets();
-		long[] presetPrices = new long[presets.Count()];
+		double[] presetPrices = new double[presets.Count()];
 		for (int j = 0; j < presetPrices.Length; j++)
 			presetPrices[j] = presets.ElementAt(j).Price;
 		//저장
